4520 removing getRecipesQuery
